💫 animation on menu

+++ src/apps/FileExplorer/Menu.svelte
@@ -2,9 +2,10 @@
   export let contextfile: string = ''
   export let files: Promise<string[]> = Promise.resolve([])
   import fs from '../../fs'
+  import { slide } from 'svelte/transition'
   export let context: HTMLMenuElement
 
-  async function download () {
+  async function download() {
     const blob = new Blob([await fs.promises.readFile(contextfile)])
     const link = document.createElement('a')
     const href = URL.createObjectURL(blob)
@@ -14,36 +15,33 @@
     URL.revokeObjectURL(href)
     contextfile = ''
   }
-  function close (e: Event) {
+  function close(e: Event) {
     if (!context.contains(e.target as Node)) contextfile = ''
   }
 </script>
+
 <svelte:body on:mousedown={close} on:touchend={close} on:keyup={close} />
-<menu
-  class:hidden={!contextfile}
-  bind:this={context}
-  on:blur={() => (contextfile = '')}
-  on:contextmenu|preventDefault
->
-  <button
-    on:click={async () => {
-      await fs.promises.unlink(contextfile)
-      files = Promise.resolve(
-        (await files).filter((file) => file !== contextfile.split('/').pop())
-      )
-      contextfile = ''
-    }}>Delete</button
+{#if contextfile}
+  <menu
+    transition:slide={{ duration: 500 }}
+    bind:this={context}
+    on:blur={() => (contextfile = '')}
+    on:contextmenu|preventDefault
   >
-  <button
-    on:click={download}>Download</button
-  >
-</menu>
+    <button
+      on:click={async () => {
+        await fs.promises.unlink(contextfile)
+        files = Promise.resolve(
+          (await files).filter((file) => file !== contextfile.split('/').pop())
+        )
+        contextfile = ''
+      }}>Delete</button
+    >
+    <button on:click={download}>Download</button>
+  </menu>
+{/if}
 
 <style>
-  .hidden {
-    display: none;
-  }
-
   menu {
     position: fixed;
     width: 200px;
@@ -53,7 +51,7 @@
   }
 
   button:hover {
-    background: rgba(255,255,255,0.1);
+    background: rgba(255, 255, 255, 0.1);
   }
   button {
     background: none;
@@ -64,8 +62,8 @@
   }
   @media (prefers-color-scheme: light) {
     button:hover {
-    background: rgba(0,0,0,0.1);
-  }
+      background: rgba(0, 0, 0, 0.1);
+    }
     button {
       color: black;
     }
